lib/boot.mk: do not define BOOTLOADER type for e2k
It is unnecessary to define the BOOTLOADER type for e2k in this context.
The reason is that images for e2k use the mkimage isodata pack method,
which is unaffected by any value of the BOOTLOADER variable (BOOT_TYPE
in the mkimage).  Therefore, defining the BOOTLOADER type does not make
any sense.

Fixes: e8fcdaf8 ("lib/boot.mk: Set IMAGE_PACKTYPE to isodata for e2k")