e2k: generate.mk fixes and tweaks
This is a cumulative patch starting with
a typo fix and ending with intentation fix,
but there's a missing "-n" bugfix in the middle
along with assorted stylistic improvements
(at least to me :).

"condition && action" lines tend to be less readable
when conditions and/or actions become more than a few
characters (especially when these come in packs);

condition &&
	action

form is less prone to some forms of human errors, and

if condition; then
	action
fi

is a guard against "just add one more action"
which gets added *unconditionally* often when
the original construct was meant as a one-liner
with strictly one action.

Making code fit 80 columns helps fit more tall xterms
horizontally when merging that code, and less than four
tends to be harmful for merge performance (takes two
FullHD displays or a 4K one for me these days).