regular.mk: refactor regular-jeos
Now these are live images instead of altinst, which have almost
nothing cut out. It is possible to use the rw-mode of live, which
can be useful for experiments and debugging the installer.
Added firmware-linux both to live and to the installed system to
expand the range of platforms on which the system can be installed.
Added the drm module to the installed system so that it is possible
most video cards support.
Now the installer can be connected via vnc.
A standard set of installer steps is used, as in other regular.
Such jeos images can be used as a base for both a desktop and a
server. This is exactly what users expect from jeos, but it was so
cut that it could not be installed on real hardware. This has now
been fixed.