Remove +systemd-optimal everything
This set of systemd settings does not seem so optimal anymore.
For example, systemd-settings-enable-showstatus breaks the output
of messages when using plymouth. On top of the messages that
plymouth has accumulated, systemd messages are superimposed.
This is critical when running the installer in systemd.
And the message wall on tty12 outputs localized messages to the
non-unicode console.