stage2: create directory /root/.kbd
When logging in, the script /etc/profile.d/console.sh is executed,
which run /usr/share/console-scripts/vt_activate_unicode.
Which in turn run /bin/unicode_start.
The directory /root/.kbd is created in /bin/unicode_start.
There is a race between tty2 and tty3, who will create the directory
/root/.kbd faster.

Solution:
Create the directory /root/.kbd in advance during building, so that
the unicode_start script does not stumble when running in parallel.

Link: https://bugzilla.altlinux.org/53223