e2k: use RESCUE_BOOTARGS for rescue instead of STAGE2_BOOTARGS
A different set of boot arguments should be used for Rescue.
But this was not done in e2k before, because on the one hand
it was easier in the case of the generator, and on the other
hand there was no need for it as plymouth was disabled there.
The need arose after enabling plymouth on e2k and combining
altinst, live and rescue functionality in one squashfs.
We need to turn off plymouth for rescue images with
'nosplash' boot argument.