net: add use/net/networkd/live
In live, need to generate network interfaces and for some cases
(installer and rescue) bring up the network. In this case, one
network interface and a timeout of 15 seconds are enough.
Also, when booting via make-initrd, problems arise with
systemd-network-generator, since they use the same arguments in
the kernel cmdline. Therefore, systemd-networkd-generator should
be disabled in live.